+++ Code/CygSoft.SmartSession.Domain/RecordingRoutines/IPracticeTimeProgress.cs
@@ -11,7 +11,7 @@ namespace CygSoft.SmartSession.Domain.RecordingRoutines
         int CurrentTime { get; }
         int TargetTime { get; }
 
-        double Weighting { get; }
+        int Weighting { get; }
         double CalculateProgress();
 
         void Increase(int value);
